pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novelscope"
version = "0.1.0"
description = "Analyze novels: chapters, metadata, tags, and more."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{ name = "You" }]
dependencies = [
  "spacy>=3.7.0",
  "pymupdf>=1.21.0",
  "pdfminer.six>=20221105",
  "python-docx>=0.8.11",
  "pydantic>=2.6.0",
  "typer[all]>=0.9.0",
  "rich>=13.0.0",
  "textstat>=0.7.3",
  "yake>=0.4.8",
  "nltk>=3.8.1",
  "sumy>=0.11.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
extras = [
  "networkx>=3.2.0",
  "keybert>=0.8.4",
  "sentence-transformers>=2.2.2",
  "hdbscan>=0.8.33",
  "bertopic>=0.16.0",
  "booknlp>=1.0.0",
  "torch>=2.2.0",
]

[project.scripts]
novelscope = "novelscope.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
